+++ providers/universal-provider/src/utils/index.ts
@@ -1,3 +1,5 @@
 export * from "./misc";
 export * from "./globals";
 export * from "./caip25";
+export * from "./storage";
+export * from "./eip5792";